+++ client/app/components/index.ts
@@ -2,3 +2,4 @@ export * from './header.component'
 export * from './task-list.component'
 export * from './task/task.component'
 export * from './create-task-list-form.component'
+export * from './create-task-form.component'
